Extract indexes in MongoConfigurationIndexStore
